Remove unnecessary array pc recording

+++ hotspot/src/share/vm/concolic/reference/array/arrayInternal.cpp
@@ -6,14 +6,14 @@
 
 void ArrayInternal::store(int index, Expression *exp) {
   assert(index >= 0, "Must be >= 0");
-  if (index >= _element_exprs.size()) {
+  if (index >= (int) _element_exprs.size()) {
     _element_exprs.resize(index);
   }
   _element_exprs[index] = exp;
 }
 
 Expression *ArrayInternal::load(int index) {
-  if (index >= _element_exprs.size()) {
+  if (index >= (int) _element_exprs.size()) {
     return NULL;
   }
